+++ e2e/connect_test.go
@@ -3,33 +3,31 @@ package dsock_test
 import (
 	"encoding/json"
 	"github.com/gorilla/websocket"
+	"github.com/stretchr/testify/suite"
 	"io/ioutil"
 	"testing"
 )
 
-func TestClaimConnect(t *testing.T) {
+type ConnectSuite struct {
+	suite.Suite
+}
+
+func TestConnectSuite(t *testing.T) {
+	suite.Run(t, new(ConnectSuite))
+}
+
+func (suite *ConnectSuite) TestClaimConnect() {
 	claim, err := createClaim(claimOptions{
 		User:    "connect",
 		Session: "claim",
 	})
-
-	if err != nil {
-		t.Fatal("Error during claim creation:", err.Error())
-	}
-
-	if claim["success"].(bool) != true {
-		t.Fatal("Application error during claim creation:", claim["error"])
+	if !checkRequestError(suite.Suite, err, claim, "claim creation") {
+		return
 	}
 
 	conn, resp, err := websocket.DefaultDialer.Dial("ws://worker/connect?claim="+claim["claim"].(map[string]interface{})["id"].(string), nil)
-	if err != nil {
-		t.Error("Error during connection ("+resp.Status+"):", err.Error())
-		body, err := ioutil.ReadAll(resp.Body)
-		if err == nil {
-			t.Fatal("Body:", string(body))
-		} else {
-			t.Fatal("Could not read body")
-		}
+	if !checkConnectionError(suite.Suite, err, resp) {
+		return
 	}
 
 	defer conn.Close()
@@ -38,60 +36,53 @@ func TestClaimConnect(t *testing.T) {
 		User:    "connect",
 		Session: "claim",
 	})
-
-	if err != nil {
-		t.Fatal("Error during getting info:", err.Error())
-	}
-
-	if info["success"].(bool) != true {
-		t.Fatal("Application error during getting info:", info["error"])
+	if !checkRequestError(suite.Suite, err, info, "getting info") {
+		return
 	}
 
 	connections := info["connections"].([]interface{})
-
-	if len(connections) != 1 {
-		t.Fatal("Invalid number of connections, expected 1:", len(connections))
+	if !suite.Len(connections, 1, "Invalid number of connections") {
+		return
 	}
 
+	claimData := claim["claim"].(map[string]interface{})
 	connection := connections[0].(map[string]interface{})
 
-	if connection["user"] != "connect" {
-		t.Fatal("Connection user did not match (expected != actual): connect !=", connection["user"])
-	}
+	suite.Equal("connect", claimData["user"], "Invalid claim user")
+	suite.Equal("connect", connection["user"], "Invalid connection user")
 
-	if connection["session"] != "claim" {
-		t.Fatal("Connection session did not match (expected != actual): jwt !=", connection["session"])
-	}
+	suite.Equal("claim", claimData["session"], "Invalid claim user session")
+	suite.Equal("claim", connection["session"], "Invalid connection user session")
 }
 
-func TestInvalidClaim(t *testing.T) {
+func (suite *ConnectSuite) TestInvalidClaim() {
 	_, resp, err := websocket.DefaultDialer.Dial("ws://worker/connect?claim=invalid-claim", nil)
-	if err == nil {
-		t.Error("Did not error when error is expected during connection")
+	if !suite.Error(err, "Did not error when expected during connection") {
+		return
 	}
 
 	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		t.Fatal("Could not read body")
+	if !suite.NoError(err, "Could not read body") {
+		return
 	}
 
 	var parsedBody map[string]interface{}
 
 	err = json.Unmarshal(body, &parsedBody)
-	if err != nil {
-		t.Fatal("Could not parse body")
+	if !suite.NoError(err, "Could not parse body") {
+		return
 	}
 
-	if parsedBody["success"] == true {
-		t.Fatal("Body said successful when should have failed")
+	if !suite.Equal(false, parsedBody["success"], "Succeeded when should have failed") {
+		return
 	}
 
-	if parsedBody["errorCode"] != "MISSING_CLAIM" {
-		t.Fatal("Incorrect error (expected != actual): MISSING_CLAIM != ", parsedBody["errorCode"])
+	if !suite.Equal("MISSING_CLAIM", parsedBody["errorCode"], "Invalid error code") {
+		return
 	}
 }
 
-func TestJwtConnect(t *testing.T) {
+func (suite *ConnectSuite) TestJwtConnect() {
 	// Hard coded JWT with max expiry:
 	// {
 	//  "sub": "connect",
@@ -101,14 +92,8 @@ func TestJwtConnect(t *testing.T) {
 	jwt := "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjb25uZWN0Iiwic2lkIjoiand0IiwiZXhwIjoyMTQ3NDg1NTQ2fQ.oMbgPfg86I1sWs6IK25AP0H4ftzUVt9asKr9W9binW0"
 
 	conn, resp, err := websocket.DefaultDialer.Dial("ws://worker/connect?jwt="+jwt, nil)
-	if err != nil {
-		t.Error("Error during connection ("+resp.Status+"):", err.Error())
-		body, err := ioutil.ReadAll(resp.Body)
-		if err == nil {
-			t.Fatal("Body:", string(body))
-		} else {
-			t.Fatal("Could not read body")
-		}
+	if !checkConnectionError(suite.Suite, err, resp) {
+		return
 	}
 
 	defer conn.Close()
@@ -117,33 +102,22 @@ func TestJwtConnect(t *testing.T) {
 		User:    "connect",
 		Session: "jwt",
 	})
-
-	if err != nil {
-		t.Fatal("Error during getting info:", err.Error())
-	}
-
-	if info["success"].(bool) != true {
-		t.Fatal("Application error during getting info:", info["error"])
+	if !checkRequestError(suite.Suite, err, info, "getting info") {
+		return
 	}
 
 	connections := info["connections"].([]interface{})
-
-	if len(connections) != 1 {
-		t.Fatal("Invalid number of connections, expected 1:", len(connections))
+	if !suite.Len(connections, 1, "Invalid number of connections") {
+		return
 	}
 
 	connection := connections[0].(map[string]interface{})
 
-	if connection["user"] != "connect" {
-		t.Fatal("Connection user did not match (expected != actual): connect !=", connection["user"])
-	}
-
-	if connection["session"] != "jwt" {
-		t.Fatal("Connection session did not match (expected != actual): jwt !=", connection["session"])
-	}
+	suite.Equal("connect", connection["user"], "Invalid connection user")
+	suite.Equal("jwt", connection["session"], "Invalid connection user session")
 }
 
-func TestInvalidJwt(t *testing.T) {
+func (suite *ConnectSuite) TestInvalidJwt() {
 	// Hard coded JWT with invalid expiry:
 	// {
 	//  "sub": "connect",
@@ -153,27 +127,27 @@ func TestInvalidJwt(t *testing.T) {
 	jwt := "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjb25uZWN0Iiwic2lkIjoiaW52YWxpZCIsImV4cCI6ImludmFsaWQifQ.afZ4Mi-K0FeS35n7sivpNlq41JUi-QKVEjkH6mGWOrk"
 
 	_, resp, err := websocket.DefaultDialer.Dial("ws://worker/connect?jwt="+jwt, nil)
-	if err == nil {
-		t.Error("Did not error when error is expected during connection")
+	if !suite.Error(err, "Did not error when expecting during connection") {
+		return
 	}
 
 	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		t.Fatal("Could not read body")
+	if !suite.NoError(err, "Could not read body") {
+		return
 	}
 
 	var parsedBody map[string]interface{}
 
 	err = json.Unmarshal(body, &parsedBody)
-	if err != nil {
-		t.Fatal("Could not parse body")
+	if !suite.NoError(err, "Could not parse body") {
+		return
 	}
 
-	if parsedBody["success"] == true {
-		t.Fatal("Body said successful when should have failed")
+	if !suite.Equal(false, parsedBody["success"], "Application succeeded when expected to fail") {
+		return
 	}
 
-	if parsedBody["errorCode"] != "INVALID_JWT" {
-		t.Fatal("Incorrect error (expected != actual): INVALID_JWT != ", parsedBody["errorCode"])
+	if !suite.Equal("INVALID_JWT", parsedBody["errorCode"], "Incorrect error code") {
+		return
 	}
 }